% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccatSALB-package.R
\docType{package}
\name{iccatSALB-package}
\alias{iccatSALB}
\alias{iccatSALB-package}
\title{iccatSALB: South Atlantic Albacore Management Strategy Evaluation}
\description{
Operating models, management procedures, and analysis tools for the Management Strategy Evaluation (MSE) of South Atlantic albacore tuna, developed for the International Commission for the Conservation of Atlantic Tunas (ICCAT).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Blue-Matter/iccatSALB}
  \item \url{https://blue-matter.github.io/iccatSALB/}
  \item Report bugs at \url{https://github.com/Blue-Matter/iccatSALB/issues}
}

}
\author{
\strong{Maintainer}: Adrian Hordyk \email{adrian@bluematterscience.com}

Authors:
\itemize{
  \item Adrian Hordyk \email{adrian@bluematterscience.com}
}

}
\keyword{internal}
